+++ app/controllers/api-v1-controller.js
@@ -6,6 +6,7 @@ var account = require( '../models/account-model' );
 var auth = require( 'basic-auth' );
 var express = require( 'express' );
 var router = express.Router();
+var quotaErrorMessage = 'Forbidden. No quota left';
 var debug = require( 'debug' )( 'api-controller' );
 
 module.exports = function( app ) {
@@ -19,6 +20,7 @@ router
         res.redirect( 'http://apidocs.enketo.org' );
     } )
     .all( '*', authCheck )
+    .all( '*', _setQuotaUsed )
     .all( '/*/iframe', _setIframeQueryParam )
     .all( '/survey/preview*', function( req, res, next ) {
         req.webformType = 'preview';
@@ -69,7 +71,6 @@ function authCheck( req, res, next ) {
 
     account.get( server )
         .then( function( account ) {
-            debug( 'account', account );
             if ( !key || ( key !== account.key ) ) {
                 error = new Error( 'Not Allowed. Invalid API key.' );
                 error.status = 401;
@@ -78,6 +79,7 @@ function authCheck( req, res, next ) {
                     .set( 'WWW-Authenticate', 'Basic realm="Enter valid API key as user name"' );
                 next( error );
             } else {
+                req.account = account;
                 next();
             }
         } )
@@ -86,6 +88,10 @@ function authCheck( req, res, next ) {
 
 function getExistingSurvey( req, res, next ) {
     var error, body;
+
+    if ( req.account.quota < req.account.quotaUsed ) {
+        return _render( 403, quotaErrorMessage, res );
+    }
 
     return surveyModel
         .getId( {
@@ -109,10 +115,16 @@ function getNewOrExistingSurvey( req, res, next ) {
             openRosaId: req.body.form_id || req.query.form_id
         };
 
+    if ( req.account.quota < req.account.quotaUsed ) {
+        return _render( 403, quotaErrorMessage, res );
+    }
+
     return surveyModel
         .getId( survey ) // will return id only for existing && active surveys
         .then( function( id ) {
-            debug( 'id: ' + id );
+            if ( !id && req.account.quota <= req.account.quotaUsed ) {
+                return _render( 403, quotaErrorMessage, res );
+            }
             status = ( id ) ? 200 : 201;
             // even if id was found still call .set() method to update any properties
             return surveyModel.set( survey )
@@ -188,6 +200,10 @@ function getList( req, res, next ) {
 function cacheInstance( req, res, next ) {
     var error, body, survey;
 
+    if ( req.account.quota < req.account.quotaUsed ) {
+        return _render( 403, quotaErrorMessage, res );
+    }
+
     survey = {
         openRosaServer: req.body.server_url,
         openRosaId: req.body.form_id,
@@ -222,6 +238,15 @@ function removeInstance( req, res, next ) {
             }
         } )
         .catch( next );
+}
+
+function _setQuotaUsed( req, res, next ) {
+    surveyModel
+        .getNumber( req.account.openRosaServer )
+        .then( function( number ) {
+            req.account.quotaUsed = number;
+            next();
+        } );
 }
 
 function _setIframeQueryParam( req, res, next ) {